refactored Connection Factory from being tightly coupled to method, allowing injection via interface to allow unit testing
