+++ src/CarsUnlimited.CartAPI/Services/ICartService.cs
@@ -1,8 +1,7 @@
 ﻿using CarsUnlimited.CartShared.Entities;
-using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
+using RabbitMQ.Client;
 
 namespace CarsUnlimited.CartAPI.Services
 {
@@ -13,6 +12,6 @@ namespace CarsUnlimited.CartAPI.Services
         Task<bool> AddToCart(CartItem cartItem);
         Task<bool> DeleteFromCart(string sessionId, string carId);
         Task<bool> DeleteAllFromCart(string sessionId);
-        Task<bool> CompleteCart(string sessionId);
+        Task<bool> CompleteCart(string sessionId, IConnectionFactory connectionFactory);
     }
 }
